fix types

+++ include/runtime/TableBuilder.h
@@ -17,14 +17,14 @@ class ResultTable {
    public:
    std::shared_ptr<arrow::Table> get();
    //interface for generated code
-   static ResultTable* create(ExecutionContext*,VarLen32 schemaDescription);
+   static ResultTable* create(ExecutionContext*, VarLen32 schemaDescription);
    static ResultTable* merge(ThreadLocal* threadLocal);
    void addBool(bool isValid, bool value);
    void addInt8(bool isValid, int8_t);
    void addInt16(bool isValid, int16_t);
    void addInt32(bool isValid, int32_t);
    void addInt64(bool isValid, int64_t);
-   void addTFloat(bool isValid, float);
+   void addTFloat(bool isValid, _Float16);
    void addFloat32(bool isValid, float);
    void addFloat64(bool isValid, double);
    void addDecimal(bool isValid, __int128);
