+++ lib/runtime/TableBuilder.cpp
@@ -115,7 +115,7 @@ class TableBuilder {
       return lowered;
    }
    TableBuilder(std::shared_ptr<arrow::Schema> schema) : schema(schema) {
-      batchBuilder=arrow::RecordBatchBuilder::Make(lowerSchema(schema), arrow::default_memory_pool()).ValueOrDie();
+      batchBuilder = arrow::RecordBatchBuilder::Make(lowerSchema(schema), arrow::default_memory_pool()).ValueOrDie();
    }
    std::shared_ptr<arrow::RecordBatch> convertBatch(std::shared_ptr<arrow::RecordBatch> recordBatch) {
       std::vector<std::shared_ptr<arrow::ArrayData>> columnData;
@@ -127,7 +127,7 @@ class TableBuilder {
    void flushBatch() {
       if (currentBatchSize > 0) {
          std::shared_ptr<arrow::RecordBatch> recordBatch;
-         recordBatch=batchBuilder->Flush(true).ValueOrDie(); //NOLINT (clang-diagnostic-unused-result)
+         recordBatch = batchBuilder->Flush(true).ValueOrDie(); //NOLINT (clang-diagnostic-unused-result)
          currentBatchSize = 0;
          batches.push_back(convertBatch(recordBatch));
       }
@@ -153,7 +153,7 @@ class TableBuilder {
    void addInt16(bool isValid, int16_t);
    void addInt32(bool isValid, int32_t);
    void addInt64(bool isValid, int64_t);
-   void addTFloat(bool isValid, int);
+   void addTFloat(bool isValid, _Float16);
    void addFloat32(bool isValid, float);
    void addFloat64(bool isValid, double);
    void addDecimal(bool isValid, __int128);
@@ -191,7 +191,7 @@ void TableBuilder::addBool(bool isValid, bool value) {
    }
 }
 
-void TableBuilder::addTFloat(bool isValid, int value) {
+void TableBuilder::addTFloat(bool isValid, _Float16 value) {
    auto* typedBuilder = getBuilder<arrow::NumericBuilder<arrow::HalfFloatType>>();
    if (!isValid) {
       handleStatus(typedBuilder->AppendNull());
@@ -264,7 +264,7 @@ RESULT_TABLE_FORWARD(addInt8, int8_t);
 RESULT_TABLE_FORWARD(addInt16, int16_t);
 RESULT_TABLE_FORWARD(addInt32, int32_t);
 RESULT_TABLE_FORWARD(addInt64, int64_t);
-RESULT_TABLE_FORWARD(addTFloat, float);
+RESULT_TABLE_FORWARD(addTFloat, _Float16);
 RESULT_TABLE_FORWARD(addFloat32, float);
 RESULT_TABLE_FORWARD(addFloat64, double);
 RESULT_TABLE_FORWARD(addDecimal, __int128);
